+++ controllers/validators/schema/user-tags.js
@@ -1,0 +1,46 @@
+'use strict';
+
+const { tagname, story, relevance } = require('./paths');
+
+const postUserTags = {
+  id: 'postUserTags',
+  properties: {
+    body: {
+      properties: {
+        tag: {
+          properties: {
+            tagname
+          }
+        },
+        story,
+        relevance
+      },
+      additionalProperties: false,
+      required: ['tag', 'story', 'relevance']
+    }
+  },
+  required: ['body']
+};
+
+const patchUserTag = {
+  id: 'patchUserTag',
+  properties: {
+    body: {
+      properties: {
+        story,
+        relevance,
+        id: {}
+      },
+      additionalProperties: false,
+      required: ['id']
+    },
+    params: {
+      properties: {
+        tagname
+      }
+    }
+  },
+  required: ['body', 'params']
+};
+
+module.exports = { postUserTags, patchUserTag };
